chore: replace accessors with accessibles, rename mixins to match official names
