--- src/main/java/com/hrznstudio/galacticraft/mixin/EmptyChunkMixin.java
@@ -1,54 +1,0 @@
-/*
- * Copyright (c) 2019-2021 HRZN LTD
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package com.hrznstudio.galacticraft.mixin;
-
-import com.hrznstudio.galacticraft.accessor.ChunkOxygenAccessor;
-import org.spongepowered.asm.mixin.Mixin;
-
-import java.util.Collections;
-import java.util.List;
-import net.minecraft.network.FriendlyByteBuf;
-import net.minecraft.network.protocol.game.ClientboundCustomPayloadPacket;
-import net.minecraft.world.level.chunk.EmptyLevelChunk;
-
-@Mixin(EmptyLevelChunk.class)
-public abstract class EmptyChunkMixin implements ChunkOxygenAccessor {
-    @Override
-    public boolean isBreathable(int x, int y, int z) {
-        return true;
-    }
-
-    @Override
-    public void setBreathable(int x, int y, int z, boolean value) {
-    }
-
-    @Override
-    public List<ClientboundCustomPayloadPacket> syncToClient() {
-        return Collections.emptyList();
-    }
-
-    @Override
-    public void readOxygenUpdate(byte b, FriendlyByteBuf packetByteBuf) {
-
-    }
-}
