--- src/main/java/com/hrznstudio/galacticraft/mixin/PlayerEntityMixin.java
@@ -1,50 +1,0 @@
-/*
- * Copyright (c) 2019-2021 HRZN LTD
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package com.hrznstudio.galacticraft.mixin;
-
-import alexiil.mc.lib.attributes.item.impl.FullFixedItemInv;
-import com.hrznstudio.galacticraft.accessor.GearInventoryProvider;
-import net.minecraft.nbt.CompoundTag;
-import net.minecraft.world.entity.player.Player;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Unique;
-
-@Mixin(Player.class)
-public abstract class PlayerEntityMixin implements GearInventoryProvider {
-    private @Unique final FullFixedItemInv gearInv = new FullFixedItemInv(12);
-
-    @Override
-    public FullFixedItemInv getGearInv() {
-        return gearInv;
-    }
-
-    @Override
-    public CompoundTag writeGearToNbt(CompoundTag tag) {
-        return getGearInv().toTag(tag);
-    }
-
-    @Override
-    public void readGearFromNbt(CompoundTag tag) {
-        getGearInv().fromTag(tag);
-    }
-}
