+++ src/main/java/com/hrznstudio/galacticraft/world/biome/source/MoonBiomeSource.java
@@ -22,7 +22,6 @@
 
 package com.hrznstudio.galacticraft.world.biome.source;
 
-import com.hrznstudio.galacticraft.mixin.BuiltinBiomesAccessor;
 import com.hrznstudio.galacticraft.world.biome.GalacticraftBiomes;
 import com.hrznstudio.galacticraft.world.biome.layer.MoonBiomeLayers;
 import com.mojang.serialization.Codec;
@@ -57,13 +56,13 @@ public class MoonBiomeSource extends BiomeSource {
         this.seed = seed;
         this.registry = registry;
         this.sampler = MoonBiomeLayers.build(seed, biomeSize, registry);
-        if (!BuiltinBiomesAccessor.getTO_NAME().containsValue(GalacticraftBiomes.Moon.HIGHLANDS_PLAINS)) {
-            BuiltinBiomesAccessor.getTO_NAME().put(registry.getId(registry.get(GalacticraftBiomes.Moon.HIGHLANDS_PLAINS)), GalacticraftBiomes.Moon.HIGHLANDS_PLAINS);
-            BuiltinBiomesAccessor.getTO_NAME().put(registry.getId(registry.get(GalacticraftBiomes.Moon.HIGHLANDS_ROCKS)), GalacticraftBiomes.Moon.HIGHLANDS_ROCKS);
-            BuiltinBiomesAccessor.getTO_NAME().put(registry.getId(registry.get(GalacticraftBiomes.Moon.HIGHLANDS_VALLEY)), GalacticraftBiomes.Moon.HIGHLANDS_VALLEY);
-            BuiltinBiomesAccessor.getTO_NAME().put(registry.getId(registry.get(GalacticraftBiomes.Moon.MARE_PLAINS)), GalacticraftBiomes.Moon.MARE_PLAINS);
-            BuiltinBiomesAccessor.getTO_NAME().put(registry.getId(registry.get(GalacticraftBiomes.Moon.MARE_ROCKS)), GalacticraftBiomes.Moon.MARE_ROCKS);
-            BuiltinBiomesAccessor.getTO_NAME().put(registry.getId(registry.get(GalacticraftBiomes.Moon.MARE_EDGE)), GalacticraftBiomes.Moon.MARE_EDGE);
+        if (!Biomes.TO_NAME.containsValue(GalacticraftBiomes.Moon.HIGHLANDS_PLAINS)) {
+            Biomes.TO_NAME.put(registry.getId(registry.get(GalacticraftBiomes.Moon.HIGHLANDS_PLAINS)), GalacticraftBiomes.Moon.HIGHLANDS_PLAINS);
+            Biomes.TO_NAME.put(registry.getId(registry.get(GalacticraftBiomes.Moon.HIGHLANDS_ROCKS)), GalacticraftBiomes.Moon.HIGHLANDS_ROCKS);
+            Biomes.TO_NAME.put(registry.getId(registry.get(GalacticraftBiomes.Moon.HIGHLANDS_VALLEY)), GalacticraftBiomes.Moon.HIGHLANDS_VALLEY);
+            Biomes.TO_NAME.put(registry.getId(registry.get(GalacticraftBiomes.Moon.MARE_PLAINS)), GalacticraftBiomes.Moon.MARE_PLAINS);
+            Biomes.TO_NAME.put(registry.getId(registry.get(GalacticraftBiomes.Moon.MARE_ROCKS)), GalacticraftBiomes.Moon.MARE_ROCKS);
+            Biomes.TO_NAME.put(registry.getId(registry.get(GalacticraftBiomes.Moon.MARE_EDGE)), GalacticraftBiomes.Moon.MARE_EDGE);
         }
     }
