+++ src/main/resources/galacticraft-rewoven.mixins.json
@@ -4,34 +4,31 @@
   "plugin": "com.hrznstudio.galacticraft.GalacticraftMixinPlugin",
   "compatibilityLevel": "JAVA_8",
   "mixins": [
-    "BaseFluidMixin",
-    "BuiltinBiomesAccessor",
+    "FlowingFluidMixin",
     "ChunkHolderMixin",
-    "ChunkSectionMixin",
+    "LevelChunkSectionMixin",
     "ChunkSerializerMixin",
     "DimensionTypeMixin",
-    "EmptyChunkMixin",
+    "EmptyLevelChunkMixin",
     "EntityMixin",
     "ItemStackMixin",
-    "LanternBlockMixin",
+    "LanternMixin",
     "LivingEntityMixin",
-    "PlayerEntityMixin",
+    "PlayerMixin",
     "ProtoChunkMixin",
-    "ServerWorldMixin",
-    "SkyPropertiesAccessor",
-    "StructurePoolGeneratorMixin",
+    "ServerLevelMixin",
+    "JigsawPlacerMixin",
     "TorchBlockMixin",
-    "WorldChunkMixin",
-    "WorldMixin"
+    "LevelChunkMixin",
+    "LevelMixin"
   ],
   "client": [
-    "client.AbstractClientPlayerEntityMixin",
-    "client.ClientChunkSectionMixin",
-    "client.HandledScreenHooks",
-    "client.InGameHudMixin",
-    "client.PauseMenuScreenMixin",
-    "client.PlayerInventoryScreenMixin",
-    "client.WorldRendererMixin"
+    "client.AbstractClientPlayerMixin",
+    "client.LevelChunkSectionMixin",
+    "client.GuiMixin",
+    "client.PauseScreenMixin",
+    "client.InventoryScreenMixin",
+    "client.LevelRendererMixin"
   ],
   "minVersion": "0.8.0",
   "injectors": {
