+++ src/main/resources/galacticraft_rewoven.accesswidener
@@ -34,6 +34,11 @@ accessible  field   net/minecraft/world/entity/npc/VillagerType             BY_B
 accessible  field   net/minecraft/world/level/levelgen/NoiseBasedChunkGenerator             BEARD_KERNEL   [F
 accessible  field   net/minecraft/world/level/levelgen/NoiseBasedChunkGenerator             BIOME_WEIGHTS   [F
 accessible  field   net/minecraft/world/item/BucketItem             content    Lnet/minecraft/world/level/material/Fluid;
+accessible  field   net/minecraft/client/renderer/DimensionSpecialEffects EFFECTS Lit/unimi/dsi/fastutil/objects/Object2ObjectMap;
+accessible  field   net/minecraft/data/worldgen/biome/Biomes TO_NAME Lit/unimi/dsi/fastutil/ints/Int2ObjectMap;
+accessible  field   net/minecraft/client/gui/screens/inventory/AbstractContainerScreen leftPos I
+accessible  field   net/minecraft/client/gui/screens/inventory/AbstractContainerScreen topPos I
+accessible  field   net/minecraft/client/gui/screens/inventory/AbstractContainerScreen imageWidth I
 
 # Classes
 accessible  class   net/minecraft/world/entity/npc/VillagerTrades$EmeraldForItems
